requirements.txt
python-dotenv~=0.19.0
loguru~=0.5.3
peewee~=3.14.8
requests~=2.26.0
pyTelegramBotAPI=3.8.2